chore: moving function to utils
