+++ src/app/components/Editor.tsx
@@ -37,7 +37,7 @@ import {
   useWarningStore,
   useYamlStore,
 } from "../lib/store";
-import { getYaml } from "../lib/utils";
+import { getYaml, collectRemovedKeys } from "../lib/utils";
 import linter from "../linter";
 import publicCodeAdapter from "../publiccode-adapter";
 import { toSemVerObject } from "../semver";
@@ -60,34 +60,6 @@ import EditorVideos from "./EditorVideos";
 import PubliccodeYmlLanguages from "./PubliccodeYmlLanguages";
 import { yamlLoadEventBus } from "./UploadPanel";
 // import EditorMDInput from "./EditorMDInput";
-
-function collectRemovedKeys(original: unknown, sanitized: unknown, prefix = ""): Array<string> {
-  const removed: Array<string> = [];
-
-  if (original && typeof original === "object" && !Array.isArray(original)) {
-    const originalObj = original as Record<string, unknown>;
-    const sanitizedObj = (sanitized && typeof sanitized === "object" && !Array.isArray(sanitized))
-      ? (sanitized as Record<string, unknown>)
-      : {};
-
-    for (const key of Object.keys(originalObj)) {
-      const nextPrefix = prefix ? `${prefix}.${key}` : key;
-      if (!(key in sanitizedObj) || (sanitizedObj as Record<string, unknown>)[key] === undefined) {
-        removed.push(nextPrefix);
-      } else {
-        removed.push(
-          ...collectRemovedKeys(
-            (originalObj as Record<string, unknown>)[key],
-            (sanitizedObj as Record<string, unknown>)[key],
-            nextPrefix
-          )
-        );
-      }
-    }
-  }
-
-  return removed;
-}
 
 const validatorFn = async (values: PublicCode) => {
   try {
